feat: move grid constants to theme.tsx
