+++ src/components/legacy/icons/header-icons/burger-menu.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { GRID_GAP_DESKTOP, GRID_GAP_MOBILE } from '../../grid/grid';
+import { GRID_GAP_DESKTOP, GRID_GAP_MOBILE } from '../../../layout/theme';
 import { up } from '../../../support/breakpoint';
 
 interface BurgerMenuProps {
